# Conflicts:
#	classes/mailster.class.php

+++ classes/manage.class.php
@@ -73,12 +73,18 @@ class MailsterManage {
 				'error_export'         => esc_html__( 'There was an error while exporting', 'mailster' ),
 				'confirm_import'       => esc_html__( 'Do you really like to import these contacts?', 'mailster' ),
 				'import_complete'      => esc_html__( 'Import complete!', 'mailster' ),
+				'choose_tags'          => esc_html__( 'Choose your tags.', 'mailster' ),
 				'confirm_delete'       => esc_html__( 'You are about to delete these subscribers permanently. This step is irreversible!', 'mailster' ) . "\n" . sprintf( esc_html__( 'Type %s to confirm deletion', 'mailster' ), '"DELETE"' ),
 				'delete_n_subscribers' => esc_html__( 'Delete %s Subscribers permanently', 'mailster' ),
 				'onbeforeunloadimport' => esc_html__( 'You are currently importing subscribers! If you leave the page all pending subscribers don\'t get imported!', 'mailster' ),
 				'onbeforeunloadexport' => esc_html__( 'Your download is preparing! If you leave this page the progress will abort!', 'mailster' ),
 			)
 		);
+
+		wp_enqueue_style( 'mailster-select2', MAILSTER_URI . 'assets/css/libs/select2' . $suffix . '.css', array(), MAILSTER_VERSION );
+		wp_enqueue_style( 'mailster-select2-theme', MAILSTER_URI . 'assets/css/select2' . $suffix . '.css', array( 'mailster-select2' ), MAILSTER_VERSION );
+		wp_enqueue_script( 'mailster-select2', MAILSTER_URI . 'assets/js/libs/select2' . $suffix . '.js', array( 'jquery' ), MAILSTER_VERSION, true );
+
 		wp_enqueue_script( 'jquery' );
 		wp_enqueue_script( 'jquery-ui-sortable' );
 		wp_enqueue_script( 'jquery-touch-punch' );
@@ -105,16 +111,16 @@ class MailsterManage {
 
 		global $wpdb;
 
-		$memory_limit       = @ini_get( 'memory_limit' );
-		$max_execution_time = @ini_get( 'max_execution_time' );
+		$memory_limit       = ini_get( 'memory_limit' );
+		$max_execution_time = ini_get( 'max_execution_time' );
 
-		@set_time_limit( 0 );
+		set_time_limit( 0 );
 
 		if ( (int) $max_execution_time < 300 ) {
-			@ini_set( 'max_execution_time', 300 );
+			ini_set( 'max_execution_time', 300 );
 		}
 		if ( (int) $memory_limit < 256 ) {
-			@ini_set( 'memory_limit', '256M' );
+			ini_set( 'memory_limit', '256M' );
 		}
 
 		if ( isset( $_FILES['async-upload'] ) ) {
@@ -137,9 +143,7 @@ class MailsterManage {
 
 			if ( ! current_user_can( 'mailster_import_subscribers' ) ) {
 
-				@header( 'Content-type: application/json' );
-				echo json_encode( $return );
-				exit;
+				wp_send_json( $return );
 			}
 
 			$raw_data = esc_textarea( stripslashes( $_POST['data'] ) );
@@ -150,9 +154,7 @@ class MailsterManage {
 
 			if ( ! current_user_can( 'mailster_import_wordpress_users' ) ) {
 
-				@header( 'Content-type: application/json' );
-				echo json_encode( $return );
-				exit;
+				wp_send_json( $return );
 			}
 
 			parse_str( $_POST['wordpressusers'], $data );
@@ -277,9 +279,7 @@ class MailsterManage {
 
 		if ( isset( $return ) ) {
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 	}
@@ -295,9 +295,7 @@ class MailsterManage {
 
 		if ( ! current_user_can( 'mailster_import_subscribers' ) ) {
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		$return['identifier'] = $identifier = $_POST['identifier'];
@@ -345,6 +343,7 @@ class MailsterManage {
 		);
 		$meta_other = array(
 			'_lists'  => esc_html__( 'Lists', 'mailster' ) . ' (' . esc_html__( 'comma separated', 'mailster' ) . ')',
+			'_tags'   => esc_html__( 'Tags', 'mailster' ) . ' (' . esc_html__( 'comma separated', 'mailster' ) . ')',
 			'_status' => esc_html__( 'Status', 'mailster' ) . ' [0...6]',
 			'_lang'   => esc_html__( 'Language', 'mailster' ),
 		);
@@ -432,9 +431,20 @@ class MailsterManage {
 				$html .= '<li><label><input name="lists[]" value="' . $list->name . '" type="checkbox"> ' . $list->name . ' <span class="count">(' . number_format_i18n( $list->subscribers ) . ')</span></label></li>';
 			}
 		}
-		$html    .= '</ul></form>';
-		$html    .= '<p><label for="new_list_name">' . esc_html__( 'Add new list', 'mailster' ) . ': </label><input type="text" id="new_list_name" value=""> <button class="button" id="addlist">' . esc_html__( 'Add', 'mailster' ) . '</button></p>
-';
+		$html .= '</ul></form>';
+		$html .= '<p><label for="new_list_name">' . esc_html__( 'Add new list', 'mailster' ) . ': </label><input type="text" id="new_list_name" value=""> <button class="button" id="addlist">' . esc_html__( 'Add', 'mailster' ) . '</button></p>';
+
+		$html .= '<h3>' . esc_html__( 'Add contacts to following tags', 'mailster' ) . ':</h3>';
+		$html .= '<form id="tags"><p>';
+		$html .= '<select multiple name="tags[]" class="tags-input">';
+		$html .= '<option></option>';
+		$tags  = mailster( 'tags' )->get();
+		foreach ( $tags as $tag ) :
+			$html .= '<option value="' . esc_attr( $tag->ID ) . '">' . esc_html( $tag->name ) . '</option>';
+		endforeach;
+		$html .= '</select>';
+		$html .= '</p></form>';
+
 		$html    .= '<h3>' . esc_html__( 'Import as', 'mailster' ) . ':</h3><p>';
 		$html    .= '<p>';
 		$statuses = mailster( 'subscribers' )->get_status( null, true );
@@ -459,9 +469,7 @@ class MailsterManage {
 
 		$return['html'] = $html;
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 
 	}
 
@@ -472,18 +480,18 @@ class MailsterManage {
 
 		define( 'MAILSTER_DO_BULKIMPORT', true );
 
-		$memory_limit       = @ini_get( 'memory_limit' );
-		$max_execution_time = @ini_get( 'max_execution_time' );
+		$memory_limit       = ini_get( 'memory_limit' );
+		$max_execution_time = ini_get( 'max_execution_time' );
 
-		@ini_set( 'display_errors', 0 );
+		ini_set( 'display_errors', 0 );
 
-		@set_time_limit( 0 );
+		set_time_limit( 0 );
 
 		if ( (int) $max_execution_time < 300 ) {
-			@ini_set( 'max_execution_time', 300 );
+			ini_set( 'max_execution_time', 300 );
 		}
 		if ( (int) $memory_limit < 256 ) {
-			@ini_set( 'memory_limit', '256M' );
+			ini_set( 'memory_limit', '256M' );
 		}
 
 		$return['success'] = false;
@@ -491,9 +499,7 @@ class MailsterManage {
 		$this->ajax_nonce( json_encode( $return ) );
 
 		if ( ! current_user_can( 'mailster_import_subscribers' ) ) {
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		$timeoffset = mailster( 'helper' )->gmt_offset( true );
@@ -510,8 +516,11 @@ class MailsterManage {
 		$order = isset( $order['order'] ) ? $order['order'] : array();
 		parse_str( $bulkdata['lists'], $lists );
 		$lists = isset( $lists['lists'] ) ? $lists['lists'] : array();
+		parse_str( $bulkdata['tags'], $tags );
+		$tags = isset( $tags['tags'] ) ? $tags['tags'] : array();
 
 		$option_list_ids = array();
+		$option_tag_ids  = array();
 
 		if ( isset( $lists ) ) {
 			foreach ( (array) $lists as $list ) {
@@ -526,6 +535,29 @@ class MailsterManage {
 				}
 
 				$option_list_ids[] = $list_id;
+			}
+		}
+
+		if ( isset( $tags ) ) {
+			foreach ( (array) $tags as $tag ) {
+
+				if ( is_numeric( $tag ) ) {
+					$tag_id = mailster( 'tags' )->get( $tag );
+
+				} else {
+					$tag_id = mailster( 'tags' )->get_by_name( $tag, 'ID' );
+				}
+
+				if ( ! $tag_id ) {
+					$tag_id = mailster( 'tags' )->add( $tag );
+					if ( is_wp_error( $tag_id ) ) {
+						continue;
+					}
+				} elseif ( isset( $tag_id->ID ) ) {
+					$tag_id = $tag_id->ID;
+				}
+
+				$option_tag_ids[] = $tag_id;
 			}
 		}
 
@@ -550,14 +582,16 @@ class MailsterManage {
 				foreach ( $raw_list as $line ) {
 
 					$list_array = array();
+					$tag_array  = array();
 					$list_ids   = $option_list_ids;
+					$tag_ids    = $option_tag_ids;
 
 					if ( ! trim( $line ) ) {
 						$bulkdata['lines']--;
 						continue;
 					}
 
-					@set_time_limit( 10 );
+					set_time_limit( 10 );
 
 					$data       = explode( $bulkdata['separator'], $line );
 					$line_count = count( $data );
@@ -608,6 +642,11 @@ class MailsterManage {
 							case '_lists':
 								$list_array = explode( ',', $d );
 								$list_array = array_map( 'trim', $list_array );
+
+								break;
+							case '_tags':
+								$tag_array = explode( ',', $d );
+								$tag_array = array_map( 'trim', $tag_array );
 
 								break;
 							case '_ip_all':
@@ -735,6 +774,35 @@ class MailsterManage {
 							mailster( 'subscribers' )->assign_lists( $subscriber_id, $list_ids, $bulkdata['existing'] == 'overwrite', $added );
 						}
 
+						foreach ( $tag_array as $tag ) {
+
+							if ( empty( $tag ) ) {
+								continue;
+							}
+
+							if ( isset( $tag_cache[ $tag ] ) ) {
+								$tag_id = $tag_cache[ $tag ];
+							} else {
+								$tag_id = mailster( 'tags' )->get_by_name( $tag, 'ID' );
+							}
+
+							if ( ! $tag_id ) {
+								$tag_id = mailster( 'tags' )->add( $tag );
+								if ( is_wp_error( $tag_id ) ) {
+									continue;
+								}
+								$tag_cache[ $tag ] = $tag_id;
+							}
+
+							$tag_ids[] = $tag_id;
+
+						}
+
+						if ( ! empty( $tag_ids ) ) {
+							$tag_ids = array_unique( $tag_ids );
+							mailster( 'subscribers' )->assign_tags( $subscriber_id, $tag_ids, $bulkdata['existing'] == 'overwrite' );
+						}
+
 						$bulkdata['imported']++;
 					}
 				}
@@ -778,9 +846,7 @@ class MailsterManage {
 		}
 		$return['success'] = true;
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 	}
 
 
@@ -794,9 +860,7 @@ class MailsterManage {
 		if ( ! current_user_can( 'mailster_export_subscribers' ) ) {
 			$return['msg'] = esc_html__( 'You are not allowed to export subscribers!', 'mailster' );
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		parse_str( $_POST['data'], $d );
@@ -864,9 +928,7 @@ class MailsterManage {
 			$return['msg'] = esc_html__( 'No Subscribers found!', 'mailster' );
 		}
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 
 	}
 
@@ -882,9 +944,7 @@ class MailsterManage {
 		if ( ! current_user_can( 'mailster_export_subscribers' ) ) {
 			$return['msg'] = esc_html__( 'You are not allowed to export subscribers!', 'mailster' );
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		$filename = get_option( 'mailster_export_filename' );
@@ -892,9 +952,7 @@ class MailsterManage {
 		if ( ! file_exists( $filename ) || ! wp_is_writable( $filename ) ) {
 			$return['msg'] = esc_html__( 'Not able to write export file', 'mailster' );
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		parse_str( $_POST['data'], $d );
@@ -942,6 +1000,9 @@ class MailsterManage {
 						break;
 					case '_listnames':
 						$val = esc_html__( 'Lists', 'mailster' );
+						break;
+					case '_tagnames':
+						$val = esc_html__( 'Tags', 'mailster' );
 						break;
 					case 'hash':
 						$val = esc_html__( 'Hash', 'mailster' );
@@ -1065,6 +1126,10 @@ class MailsterManage {
 						$list = mailster( 'subscribers' )->get_lists( $user->ID );
 						$val  = implode( ', ', wp_list_pluck( $list, 'name' ) );
 						break;
+					case '_tagnames':
+						$tag = mailster( 'subscribers' )->get_tags( $user->ID );
+						$val = implode( ', ', wp_list_pluck( $tag, 'name' ) );
+						break;
 					case 'status':
 						$val = $statusnames[ $user->status ];
 						break;
@@ -1151,7 +1216,7 @@ class MailsterManage {
 
 				foreach ( $row as $key => $r ) {
 					$type = 'String';
-					if ( is_numeric( $r ) ) {
+					if ( in_array( $key, array( 'ID', '_number', '_statuscode', 'rating', 'timeoffset' ) ) ) {
 						$type = 'Number';
 					}
 					$output .= '<mailster:Cell><mailster:Data mailster:Type="' . $type . '">' . esc_html( $r ) . '</mailster:Data></mailster:Cell>' . "\n";
@@ -1174,11 +1239,13 @@ class MailsterManage {
 			} else {
 				$return['finished'] = true;
 
-				$finalname         = MAILSTER_UPLOAD_DIR . '/mailster_export_' . date( 'Y-m-d-H-i-s' ) . '.' . $outputformat;
-				$return['success'] = copy( $filename, $finalname );
-				$file_size         = @filesize( $filename );
-				update_option( 'mailster_export_filename', $finalname );
-				@unlink( $filename );
+				$finalname = MAILSTER_UPLOAD_DIR . '/mailster_export_' . date( 'Y-m-d-H-i-s' ) . '.' . $outputformat;
+				if ( file_exists( $filename ) ) {
+					$return['success'] = copy( $filename, $finalname );
+					$file_size         = filesize( $filename );
+					update_option( 'mailster_export_filename', $finalname );
+					unlink( $filename );
+				}
 				$return['filename'] = admin_url( 'admin-ajax.php?action=mailster_download_export_file&file=' . basename( $finalname ) . '&format=' . $outputformat . '&_wpnonce=' . wp_create_nonce( 'mailster_nonce' ) );
 			}
 
@@ -1191,9 +1258,7 @@ class MailsterManage {
 
 		}
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 	}
 
 
@@ -1221,6 +1286,7 @@ class MailsterManage {
 				break;
 			case 'xls':
 				header( 'Content-Type: application/vnd.ms-excel; name="' . $filename . '"' );
+				break;
 			case 'csv':
 				header( 'Content-Type: text/csv; name="' . $filename . '"' );
 				header( 'Content-Transfer-Encoding: binary' );
@@ -1235,7 +1301,7 @@ class MailsterManage {
 		if ( 'html' == $format ) {
 			echo '<table>' . "\n";
 		} elseif ( 'xls' == $format ) {
-			echo '<?xml version="1.0"?>' . "\n";
+			echo '<?xml version="1.0" encoding="UTF-8"?>' . "\n";
 			echo '<mailster:Workbook xmlns:mailster="urn:schemas-microsoft-com:office:spreadsheet">' . "\n";
 			echo '<mailster:Styles><mailster:Style mailster:ID="1"><mailster:Font mailster:Bold="1"/></mailster:Style></mailster:Styles>' . "\n";
 			echo '<mailster:Worksheet mailster:Name="' . esc_attr__( 'Mailster Subscribers', 'mailster' ) . '">' . "\n";
@@ -1271,9 +1337,7 @@ class MailsterManage {
 		if ( ! current_user_can( 'mailster_bulk_delete_subscribers' ) ) {
 			$return['msg'] = 'no allowed';
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		parse_str( $_POST['data'], $d );
@@ -1332,9 +1396,7 @@ class MailsterManage {
 			$return['msg'] = esc_html__( 'No Subscribers removed', 'mailster' );
 		}
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 
 	}
